testing gcstats and gcinfo
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@9517 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ code/base/dso-l2/src/com/tc/objectserver/impl/GCLogger.java
@@ -25,11 +25,11 @@ public class GCLogger {
     if (verboseGC()) logGC("GC: " + (fullGC ? "Full GC" : "YoungGen GC") + " START " + iteration);
   }
 
-  public void log_markStart(int size) {
+  public void log_markStart(long size) {
     if (verboseGC()) logGC("GC: pre-GC managed id count: " + size);
   }
 
-  public void log_markResults(int size) {
+  public void log_markResults(long size) {
     if (verboseGC()) logGC("GC: pre-rescue GC results: " + size);
   }
 
@@ -41,11 +41,11 @@ public class GCLogger {
     if (verboseGC()) logGC("GC: paused.");
   }
 
-  public void log_rescue_complete(int pass, int count) {
+  public void log_rescue_complete(int pass, long count) {
     if (verboseGC()) logGC("GC: rescue pass " + pass + " completed. gc candidates = " + count + " objects...");
   }
 
-  public void log_rescue_start(int pass, int count) {
+  public void log_rescue_start(int pass, long count) {
     if (verboseGC()) logGC("GC: rescue pass " + pass + " on " + count + " objects...");
   }
